# receiver/pinepeach-Retail_Store_POS_Tracker-1.0.0-swagger.yaml
openapi: 3.0.0
info:
  description: This receives data from POS systems
  version: "1.0.0"
  title: Simple Inventory API
  contact:
    email: jasonquach1234@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: transaction
paths:
  /completed/sales:
    post:
      tags:
        - transaction
      summary: adds completed sales transaction to the system
      operationId: app.record_completed_sales
      description: records the completed sales transaction
      responses:
        '201':
          description: item created
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletedSale'
        description: Completed sale to add
        
  /completed/returns:
    post:
      tags:
        - transaction
      summary: adds completed return transation to the system
      operationId: app.record_completed_returns
      description: records the completed return transaction
      responses:
        '201':
          description: item created
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletedReturn'
        description: Completed return to add
        
components:
  schemas:
    CompletedSale:
      required:
        - title
        - gcid
        - sale_product
      properties:
        title:
          type: string
          example: Sale
        gcid:
          $ref: '#/components/schemas/GCID'
        sale_product:
          $ref: '#/components/schemas/SaleProduct'
      type: object
      
    CompletedReturn:
      required:
        - title
        - gcid
        - return_product
      properties:
        title:
          type: string
          example: Return
        gcid:
          $ref: '#/components/schemas/GCID'
        return_product:
          $ref: '#/components/schemas/ReturnProduct'
      type: object
    
    GCID:
      required:
        - store
        - till_number
        - transaction_number
        - date
      properties:
        store:
          type: integer
          example: 125
        till_number:
          type: integer
          example: 5
        transaction_number:
          type: integer
          example: 2050
        date:
          type: string
          format: mm-dd-yyyy
          example: 08-12-2020
      type: object
      
    SaleProduct:
      required: 
        - sku
        - name
        - price
      properties:
        sku:
          type: integer
          format: uid
          example: 17812516
        name:
          type: string
          example: Airpods
        price:
          type: string
          example: $119.99
      type: object
      
    ReturnProduct:
      required: 
        - sku
        - name
        - reason
      properties:
        sku:
          type: integer
          format: uid
          example: 17812516
        name:
          type: string
          example: Airpods
        reason:
          type: string
          example: broken in one ear
      type: object